various fixes to make proxy work (#24)

+++ curiefense/images/curieproxy-nginx/nginx-conf-reload.sh
@@ -1,7 +1,12 @@
 #! /bin/bash
+pid_file="/usr/local/openresty/nginx/logs/nginx.pid"
 source_file="/cf-config/current/config/customconf.tar.gz"
 target_dir="/etc/nginx/conf.d"
 temp_dir="/tmp/current-conf"
+if [ ! -f "$source_file" ]; then
+    echo "$source_file does not exist."
+    exit 0
+fi
 mkdir -p ${temp_dir}
 [ ! -d "/lua/customcode" ] && mkdir -p /lua/customcode
 rm -rf ${temp_dir}/*
@@ -22,6 +27,19 @@ then
   cp ${targer_dir}/lua/customcode.lua /lua/customcode/lua.lua
 else
   ## reload nginx
-  nginx -s reload
-  echo "Nginx reloaded with the new config"
+  n=0
+  until [ "$n" -ge 10 ]
+  do
+    p_id=$(cat $pid_file)
+    [ -n "$p_id" ]  && [ "$p_id" -gt 0 ] && echo "NGINX started" && break
+    echo nginx not started
+    n=$((n+1))
+    sleep 1
+  done
+  if  [ -e $pid_file ] && p_id=$(cat $pid_file) && [ -n "$p_id" ]  && [ "$p_id" -gt 0 ] && ps "$p_id" > /dev/null
+  then
+    echo reloading nginx
+    nginx -s reload
+    echo "Nginx reloaded with the new config"
+  fi
 fi
